Folded Fleet/Corps into Military Model / updateStats on load
