+++ server/wts/json/facilities.json
@@ -4,7 +4,7 @@
 "code": "CF-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "City",
+"unitType": ["City"],
 "desc": "Civilian Improvment Facility",
 "capacity": 2,
 "effect": [],
@@ -19,7 +19,7 @@
 "code": "MF-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "City",
+"unitType": ["City"],
 "desc": "Civilian Improvment Facility",
 "capacity": 2,
 "effect": [{ "type": "pr", "effect": 1}],
@@ -34,7 +34,7 @@
 "code": "CC-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Crisis",
+"unitType": ["Crisis"],
 "desc": "Mobile Crisis Control Facility",
 "capacity": 2,
 "effect": [],
@@ -49,7 +49,7 @@
 "code": "BL-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Base",
+"unitType": ["Base"],
 "desc": "Standard Base Lab Facility",
 "capacity": 1,
 "effect": [],
@@ -67,7 +67,7 @@
 "code": "BH-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Base",
+"unitType": ["Base"],
 "desc": "Standard Base Hanger Facility",
 "capacity": 5,
 "effect": [],
@@ -82,7 +82,7 @@
 "code": "BSL-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Base",
+"unitType": ["Base"],
 "desc": "Secret Base Lab Facility",
 "capacity": 1,
 "effect": [],
@@ -100,7 +100,7 @@
 "code": "BF-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Base",
+"unitType": ["Base"],
 "desc": "Base Factory Facility",
 "capacity": 1,
 "effect": [],
@@ -115,7 +115,7 @@
 "code": "SL-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Spacecraft",
+"unitType": ["Spacecraft"],
 "desc": "Standard Spacecraft Lab Facility",
 "capacity": 1,
 "effect": [],
@@ -133,7 +133,7 @@
 "code": "SH-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Spacecraft",
+"unitType": ["Spacecraft"],
 "desc": "Standard Spacecraft Hanger Facility",
 "capacity": 3,
 "effect": [],
@@ -148,7 +148,7 @@
 "code": "CSL-1",
 "cost": 2,
 "buildtime": 1,
-"unitType": "Crash",
+"unitType": ["Crash"],
 "desc": "Mobile Research Lab Facility",
 "capacity": 1,
 "effect": [],
